Add wait/polling loop for deleteInput lambda
Background
=============
When deleting the stack, the custom resource lambda calls functions to clean up the resources.
It deletes the MediaLive channel first before deleting the MediaLive input because the input needs to be in DETACHED state in order to be deleted successfully.
For some reason with this sdk update, the input doesn't get to DETACHED state before the cleanup function is called, so the stack deletion fails.

Changes
=======
Added a wait/polling loop for the deleteInput custom resource function to make sure the MediaLive input is in DETACHED state before calling DeleteInput command.